requirements.txt
Flask==3.0.0
Flask-MySQL==2.0.0

